renamed _helpers.hpp to hashdex.hpp
