+++ pp/entrypoint/wal_decoder.cpp
@@ -2,8 +2,8 @@
 
 #include <cstdint>
 
-#include "_helpers.hpp"
 #include "exception.hpp"
+#include "hashdex.hpp"
 #include "head/lss.h"
 #include "primitives/go_slice.h"
 #include "primitives/go_slice_protozero.h"
