+++ pp/entrypoint/wal_encoder.cpp
@@ -2,8 +2,8 @@
 
 #include <variant>
 
-#include "_helpers.hpp"
 #include "exception.hpp"
+#include "hashdex.hpp"
 #include "primitives/go_slice.h"
 #include "wal/encoder.h"
 #include "wal/wal.h"
